Removing unused things.
