--- src/prosemirror-plugins/TimestampList.jsx
@@ -1,21 +1,0 @@
-/*
-
-import React, { Component, useState } from 'react';
-import MultiSelect from '@kenshooui/react-multi-select';
-
-const AddToTimestampListBtn = ({ getCurrentTime }) => {};
-
-const TimestampList = ({ timestamps, handleOnSelect }) => {
-  const [selectedItems, setSelectedItems] = useState([]);
-
-  if (!handleOnSelect) {
-    handleOnSelect = newSelectedItems => {
-      console.log('Selected items =', items);
-      setSelectedItems(newSelectedItems);
-    };
-  }
-
-  return <MultiSelect items={timestamps} />;
-};
-
-*/
